Small improvements
Files reorganization;
queue dispatch moved to console logger;
Appl native logger
