+++ Astaroth/Classes/Default.swift
@@ -4,15 +4,7 @@
 
 import Foundation
 
-public class StringTag : LogType {
-    public let logTag : String
-
-    public init(_ tag: String) {
-        self.logTag = tag
-    }
-}
-
-let Default: LogType = StringTag("+")
+public let Default: LogType = StringTag("+")
 
 public extension Logger {
     func v(_ message: @autoclosure @escaping () -> Any, file: String = #file, method: String = #function, line: UInt = #line, column: UInt = #column) {
