+++ Astaroth/Classes/LogLevel.swift
@@ -1,0 +1,12 @@
+//
+//  LogLevel.swift
+//  
+//
+//  Created by Алексей Лысенко on 26.09.2023.
+//
+
+import Foundation
+
+public enum LogLevel: Int {
+    case verbose, debug, info, warning, error, whatTheFuck
+}
