+++ Astaroth/Classes/Loggers/ConsoleLogger.swift
@@ -6,18 +6,12 @@ import Foundation
 
 func errorLevelSign(_ level: LogLevel) -> String {
     switch level {
-    case .verbose:
-        return "⚪"
-    case .debug:
-        return "✅"
-    case .info:
-        return "🔵"
-    case .warning:
-        return "⚠️"
-    case .error:
-        return "🆘"
-    case .whatTheFuck:
-        return "⁉️"
+    case .verbose: return "⚪"
+    case .debug: return "✅"
+    case .info: return "🔵"
+    case .warning: return "⚠️"
+    case .error: return "🆘"
+    case .whatTheFuck: return "⁉️"
     }
 }
 
@@ -30,13 +24,17 @@ public class ConsoleLogger: Logger {
 
     public var config: LoggerConfig
 
+    private let queue = DispatchQueue(label: "Log", qos: .default)
+
     public func log(_ lazyMessage: Lazy<Any>, level: LogLevel, type: LogType, _ context: StackTraceElement?) {
         if !isAbleToLog(level: level, type: type) { return }
 
         let tag = type.logTag
         let systemLevel = levelSign(level)
         let message = config.format(lazyMessage.value, context)
-
-        print("\(systemLevel)[\(tag)]\(message)\(systemLevel)")
+        
+        queue.async {
+            print("\(systemLevel)[\(tag)]\(message)\(systemLevel)")
+        }
     }
 }
