+++ README.md
@@ -1,9 +1,9 @@
 # Astaroth
 
-[![CI Status](https://img.shields.io/travis/horovodovodo4ka/Astaroth.svg?style=flat)](https://travis-ci.org/horovodovodo4ka/Astaroth)
 [![Version](https://img.shields.io/cocoapods/v/Astaroth.svg?style=flat)](https://cocoapods.org/pods/Astaroth)
 [![License](https://img.shields.io/cocoapods/l/Astaroth.svg?style=flat)](https://cocoapods.org/pods/Astaroth)
 [![Platform](https://img.shields.io/cocoapods/p/Astaroth.svg?style=flat)](https://cocoapods.org/pods/Astaroth)
+[![Swift Package Manager](https://img.shields.io/badge/Swift_Package_Manager-compatible-orange?style=flat-square)](https://img.shields.io/badge/Swift_Package_Manager-compatible-orange?style=flat-square)
 
 ## Example
 
@@ -13,6 +13,7 @@ To run the example project, clone the repo, and run `pod install` from the Examp
 
 ## Installation
 
+### Cocoapods
 Astaroth is available through [CocoaPods](https://cocoapods.org). To install
 it, simply add the following line to your Podfile:
 
@@ -20,9 +21,24 @@ it, simply add the following line to your Podfile:
 pod 'Astaroth'
 ```
 
+### Swift Package Manager
+The [Swift Package Manager](https://swift.org/package-manager/) is a tool for automating the distribution of Swift code and is integrated into the `swift` compiler.
+
+Once you have your Swift package set up, adding Alamofire as a dependency is as easy as adding it to the `dependencies` value of your `Package.swift`.
+
+```swift
+dependencies: [
+    .package(url: "https://github.com/horovodovodo4ka/astaroth-ios.git", .upToNextMajor(from: "0.6.0"))
+]
+```
+
 ## Author
 
 horovodovodo4ka, xbitstream@gmail.com
+
+## Contributors
+
+abesmon, abesmon@gmail.com
 
 ## License
